How It Works
since most companies that i've emailed regarding a bug in their website since im a member of it, are not taking us seriously, can we do this email aliases a more functional one, to add sending capabilities, for example we can use this in my case weblord@wearehackerone.com as a sending email to the prospective company who is not listed in the directory that they have a bug in their website and of course we have our email sig showing our hackerone.com/weblord profile so they can understand that it is their security that we're thinking of not to hack them.  make it #4 on ###how it works.